Moving unit tests to new code
