+++ app/styles/common-buttons.scss
@@ -38,7 +38,7 @@
   background-color: $delete-button-color;
   border-color: $delete-button-color;
   border-radius: $button-radius-circle;
-  padding: 4px 1px 0 2px;
+  padding: 4px 1px 0 1.5px;
   color: $white;
   height: 24px;
   width: 24px;
