+++ app/styles/components/popovers/info-popover.scss
@@ -31,6 +31,7 @@
       font-weight: $font-weight-regular;
       color: $info-text-color;
       padding: 10px 15px 0 0;
+      scrollbar-width: thin;
 
       .content-header {
         font-weight: $font-weight-strong;
