+++ tests/fixtures/queries.js
@@ -16,7 +16,7 @@ const QUERIES = {
   AND_LIST_TUMBLING_WINDOW: {
     filters: [FILTERS.AND_LIST],
     window: {
-      emit: { type: 'TIME', every: 2000 }
+      emitType: 'TIME', emitEvery: 2000
     },
     duration: 20000
   }
