+++ tests/unit/utils/window-cache-test.js
@@ -1,0 +1,11 @@
+import windowCache from 'bullet-ui/utils/window-cache';
+import { module, test } from 'qunit';
+
+module('Unit | Utility | window-cache', function() {
+
+  // Replace this with your real tests.
+  test('it works', function(assert) {
+    let result = windowCache();
+    assert.ok(result);
+  });
+});
