+++ tests/unit/validators/aggregation-max-size-test.js
@@ -7,13 +7,14 @@ import EmberObject from '@ember/object';
 import { module, test } from 'qunit';
 import { setupTest } from 'ember-qunit';
 import { AGGREGATIONS } from 'bullet-ui/models/aggregation';
+import validateAggregationMaxSize from 'bullet-ui/validators/aggregation-max-size';
 
-module('Unit | Validator | aggregation-max-size', function(hooks) {
+module('Unit | Validator | aggregation max size', function(hooks) {
   setupTest(hooks);
 
   test('it does not let you exceed the maximum aggregation size', function(assert) {
-    var validator = this.owner.lookup('validator:aggregation-max-size');
-    let mockModel = EmberObject.create({
+    var validate = validateAggregationMaxSize();
+    let mockChanges = EmberObject.create({
       type: AGGREGATIONS.get('GROUP'),
       settings: {
         defaultValues: {
@@ -23,12 +24,12 @@ module('Unit | Validator | aggregation-max-size', function(hooks) {
       }
     });
     let expected = 'The maintainer has configured Bullet to support a maximum of 500 for result count';
-    assert.equal(validator.validate(501, null, mockModel), expected);
-    assert.ok(validator.validate(500, null, mockModel));
+    assert.equal(validate('size', 501, 20, undefined, mockModel), expected);
+    assert.ok(validate('size', 500, 20, undefined, mockModel));
   });
 
   test('it does not let you exceed the maximum size for the Raw aggregation', function(assert) {
-    var validator = this.owner.lookup('validator:aggregation-max-size');
+    var validate = validateAggregationMaxSize();
     let mockModel = EmberObject.create({
       type: AGGREGATIONS.get('RAW'),
       settings: {
@@ -39,7 +40,7 @@ module('Unit | Validator | aggregation-max-size', function(hooks) {
       }
     });
     let expected = 'The maintainer has set the Raw type to support a maximum of 100 for result count';
-    assert.equal(validator.validate(101, null, mockModel), expected);
-    assert.ok(validator.validate(100, null, mockModel));
+    assert.equal(validate('size', 101, 20, undefined, mockModel), expected);
+    assert.ok(validate('size', 100, 20, undefined, mockModel));
   });
 });
