+++ tests/unit/validators/metric-field-test.js
@@ -7,28 +7,30 @@ import EmberObject from '@ember/object';
 import { module, test } from 'qunit';
 import { setupTest } from 'ember-qunit';
 import { METRICS } from 'bullet-ui/models/metric';
+import validateMetricField from 'bullet-ui/validators/metric-field';
 
-module('Unit | Validator | metric-field', function(hooks) {
+module('Unit | Validator | metric field', function(hooks) {
   setupTest(hooks);
 
   test('it checks to see if all metrics besides count have a field', function(assert) {
-    var validator = this.owner.lookup('validator:metric-field');
+    let validate = validateMetricField();
     let mockModel = EmberObject.create({
       type: METRICS.get('SUM')
     });
     let expected = `All metrics but ${METRICS.get('COUNT')} require a field`;
-    assert.equal(validator.validate(null, null, mockModel), expected);
+    assert.equal(validate('field', null, null, mockModel, undefined), expected);
 
     mockModel.set('type', METRICS.get('MIN'));
-    assert.equal(validator.validate(null, null, mockModel), expected);
+    assert.equal(validate('field', null, null, undefined, mockModel), expected);
 
     mockModel.set('type', METRICS.get('AVG'));
-    assert.equal(validator.validate(null, null, mockModel), expected);
+    assert.equal(validate('field', null, null, mockModel, undefined), expected);
 
-    mockModel.set('type', METRICS.get('MAX'));
-    assert.equal(validator.validate(null, null, mockModel), expected);
+    mockModel.set('field', METRICS.get('MAX'));
+    assert.equal(validate('field', null, null, undefined, mockModel), expected);
 
     mockModel.set('type', METRICS.get('COUNT'));
-    assert.ok(validator.validate(null, null, mockModel));
+    assert.ok(validator.validate('type', null, null, mockModel, undefined));
+    assert.ok(validator.validate('type', null, null, undefined, mockModel));
   });
 });
