+++ tests/unit/validators/query-max-duration-test.js
@@ -6,13 +6,15 @@
 import EmberObject from '@ember/object';
 import { module, test } from 'qunit';
 import { setupTest } from 'ember-qunit';
+import validateQueryMaxDuration from 'bullet-ui/validators/query-max-duration';
 
-module('Unit | Validator | query-max-duration', function(hooks) {
+module('Unit | Validator | query max duration', function(hooks) {
   setupTest(hooks);
 
   test('it does not let you exceed the maximum duration seconds', function(assert) {
-    var validator = this.owner.lookup('validator:query-max-duration');
+    let validate = validateQueryMaxDuration();
     let mockModel = EmberObject.create({
+      duration: 120,
       settings: {
         defaultValues: {
           durationMaxSecs: 120
@@ -20,7 +22,8 @@ module('Unit | Validator | query-max-duration', function(hooks) {
       }
     });
     let expected = 'The maintainer has configured Bullet to support a maximum of 120s for maximum duration';
-    assert.equal(validator.validate(121, null, mockModel), expected);
-    assert.ok(validator.validate(120, null, mockModel));
+    assert.equal(validator.validate('duration', 121, 120, undefined, mockModel), expected);
+    mockModel.set('duration', null);
+    assert.ok(validator.validate('duration', 120, 120, undefined, mockModel));
   });
 });
