--- tests/unit/validators/query-test.js
@@ -1,8 +1,0 @@
-import { module, test } from 'qunit';
-import validateQuery from 'bullet-ui/validators/query';
-
-module('Unit | Validator | query');
-
-test('it exists', function(assert) {
-  assert.ok(validateQuery());
-});
