+++ tests/unit/validators/valid-points-test.js
@@ -6,11 +6,8 @@
 import EmberObject from '@ember/object';
 import { module, test } from 'qunit';
 import { setupTest } from 'ember-qunit';
-import {
-  AGGREGATIONS,
-  DISTRIBUTIONS,
-  DISTRIBUTION_POINTS
-} from 'bullet-ui/models/aggregation';
+import { AGGREGATIONS, DISTRIBUTIONS, DISTRIBUTION_POINTS } from 'bullet-ui/models/aggregation';
+import validatePoints from 'bullet-ui/validators/query-max-duration';
 
 module('Unit | Validator | valid-points', function(hooks) {
   setupTest(hooks);
@@ -24,17 +21,18 @@ module('Unit | Validator | valid-points', function(hooks) {
   };
 
   test('it ignores non distribution type aggregations', function(assert) {
-    var validator = this.owner.lookup('validator:valid-points');
+    let validate = validatePoints();
     let mockModel = EmberObject.create({
       type: AGGREGATIONS.get('RAW')
     });
-    assert.ok(validator.validate(null, null, mockModel));
+    assert.ok(validator.validate('type', null, null, mockModel, undefined));
+    assert.ok(validator.validate('type', null, null, undefined, mockModel));
     mockModel.set('type', AGGREGATIONS.get('COUNT_DISTINCT'));
-    assert.ok(validator.validate(null, null, mockModel));
+    assert.ok(validator.validate('type', null, null, mockModel, undefined));
   });
 
   test('it validates distribution type aggregations with a number of points', function(assert) {
-    var validator = this.owner.lookup('validator:valid-points');
+    let validate = validatePoints();
     let expected;
     let mockModel = EmberObject.create({
       type: AGGREGATIONS.get('DISTRIBUTION'),
@@ -46,22 +44,22 @@ module('Unit | Validator | valid-points', function(hooks) {
       }
     });
     expected = 'You must specify the Number of Points you want to generate';
-    assert.equal(validator.validate(null, null, mockModel), expected);
+    assert.equal(validator.validate('type', null, null, mockModel, undefined), expected);
 
     mockModel.set('attributes.numberOfPoints', 21);
     expected = 'The maintainer has set the maximum number of points you can generate to be 20';
-    assert.equal(validator.validate(null, null, mockModel), expected);
+    assert.equal(validator.validate('type', null, null, mockModel, undefined), expected);
 
     mockModel.set('attributes.numberOfPoints', -1);
     expected = 'You must specify a positive Number of Points';
-    assert.equal(validator.validate(null, null, mockModel), expected);
+    assert.equal(validator.validate('type', null, null, mockModel, undefined), expected);
 
     mockModel.set('attributes.numberOfPoints', 15);
-    assert.ok(validator.validate(null, null, mockModel));
+    assert.ok(validator.validate('type', null, null, mockModel, undefined));
   });
 
   test('it validates distribution type aggregations with generated points', function(assert) {
-    var validator = this.owner.lookup('validator:valid-points');
+    let validate = validatePoints();
     let expected;
     let mockModel = EmberObject.create({
       type: AGGREGATIONS.get('DISTRIBUTION'),
@@ -78,52 +76,52 @@ module('Unit | Validator | valid-points', function(hooks) {
     mockModel.set('attributes.start', null);
     mockModel.set('attributes.end', null);
     mockModel.set('attributes.increment', null);
-    assert.equal(validator.validate(null, null, mockModel), expected);
+    assert.equal(validator.validate('attributes.start', null, null, mockModel, undefined), expected);
 
     mockModel.set('attributes.start', 15);
-    assert.equal(validator.validate(null, null, mockModel), expected);
+    assert.equal(validator.validate('attributes.start', null, null, mockModel, undefined), expected);
 
     mockModel.set('attributes.end', 500);
-    assert.equal(validator.validate(null, null, mockModel), expected);
+    assert.equal(validator.validate('attributes.start', null, null, mockModel, undefined), expected);
 
     mockModel.set('attributes.increment', 500);
-    assert.ok(validator.validate(null, null, mockModel));
+    assert.ok(validator.validate('attributes.start', null, null, mockModel, undefined));
 
     expected = 'You must specify Start less than End';
     mockModel.set('attributes.start', 10);
     mockModel.set('attributes.end', 1);
     mockModel.set('attributes.increment', 2);
-    assert.equal(validator.validate(null, null, mockModel), expected);
+    assert.equal(validator.validate('attributes.start', null, null, mockModel, undefined), expected);
 
     expected = 'You must specify a positive Increment';
     mockModel.set('attributes.start', 1);
     mockModel.set('attributes.end', 10);
     mockModel.set('attributes.increment', 0);
-    assert.equal(validator.validate(null, null, mockModel), expected);
+    assert.equal(validator.validate('attributes.start', null, null, mockModel, undefined), expected);
     mockModel.set('attributes.increment', -1);
-    assert.equal(validator.validate(null, null, mockModel), expected);
+    assert.equal(validator.validate('attributes.start', null, null, mockModel, undefined), expected);
 
     expected = 'Quantiles requires that you specify a Start and End between 0 and 1';
     mockModel.set('attributes.start', -1);
     mockModel.set('attributes.end', 1);
     mockModel.set('attributes.increment', 0.1);
     mockModel.set('attributes.type', DISTRIBUTIONS.get('QUANTILE'));
-    assert.equal(validator.validate(null, null, mockModel), expected);
+    assert.equal(validator.validate('attributes.type', null, null, mockModel, undefined), expected);
     mockModel.set('attributes.end', 10);
-    assert.equal(validator.validate(null, null, mockModel), expected);
+    assert.equal(validator.validate('attributes.end', null, null, mockModel, undefined), expected);
 
     expected = 'The maintainer has set the maximum number of points you can generate to be 20';
     mockModel.set('attributes.start', 0);
     mockModel.set('attributes.end', 1);
     mockModel.set('attributes.increment', 0.01);
-    assert.equal(validator.validate(null, null, mockModel), expected);
+    assert.equal(validator.validate('attributes.end', null, null, mockModel, undefined), expected);
 
     mockModel.set('attributes.increment', 0.05);
-    assert.ok(validator.validate(null, null, mockModel));
+    assert.ok(validator.validate('attributes.increment', null, null, mockModel, undefined));
   });
 
   test('it validates distribution type aggregations with free-form points', function(assert) {
-    var validator = this.owner.lookup('validator:valid-points');
+    let validate = validatePoints();
     let expected;
     let mockModel = EmberObject.create({
       type: AGGREGATIONS.get('DISTRIBUTION'),
@@ -135,23 +133,23 @@ module('Unit | Validator | valid-points', function(hooks) {
       }
     });
     expected = 'You must specify a comma separated list of points for this option';
-    assert.equal(validator.validate(null, null, mockModel), expected);
+    assert.equal(validator.validate('attribute.pointType' null, null, mockModel, undefined), expected);
 
     expected = 'These are not valid points: e,f';
     mockModel.set('attributes.points', '15, e, 235, 4, f, 0');
-    assert.equal(validator.validate(null, null, mockModel), expected);
+    assert.equal(validator.validate('attribute.points' null, null, mockModel, undefined), expected);
 
     expected = 'Quantiles requires points between 0 and 1. These are not: -0.4,1.2';
     mockModel.set('attributes.type', DISTRIBUTIONS.get('QUANTILE'));
     mockModel.set('attributes.points', '0.3, -0.4, 0.35, 0.4, 1.2, 0');
-    assert.equal(validator.validate(null, null, mockModel), expected);
+    assert.equal(validator.validate('attribute.points' null, null, mockModel, undefined), expected);
 
     expected = 'The maintainer has set the maximum number of points you can generate to be 5';
     mockModel.set('settings.defaultValues.sketches.distributionMaxNumberOfPoints', 5);
     mockModel.set('attributes.points', '0.3, 0.35, 0.4, 0, 0.6, 0.99');
-    assert.equal(validator.validate(null, null, mockModel), expected);
+    assert.equal(validator.validate('attribute.points' null, null, mockModel, undefined), expected);
 
     mockModel.set('attributes.points', '0.3, 0.4, 0, 0.6, 0.99');
-    assert.ok(validator.validate(null, null, mockModel));
+    assert.ok(validator.validate('attribute.points' null, null, mockModel, undefined));
   });
 });
